+++ hyp3_gamma/__main__.py
@@ -46,6 +46,7 @@ def rtc():
     parser.add_argument('--include-dem', type=string_is_true, default=False)
     parser.add_argument('--include-inc-map', type=string_is_true, default=False)
     parser.add_argument('--include-scattering-area', type=string_is_true, default=False)
+    parser.add_argument('--include-rgb', type=string_is_true, default=False)
     parser.add_argument('granule')
     args = parser.parse_args()
 
@@ -66,6 +67,7 @@ def rtc():
                         include_dem=args.include_dem,
                         include_inc_map=args.include_inc_map,
                         include_scattering_area=args.include_scattering_area,
+                        include_rgb=args.include_rgb,
                     )
     output_zip = make_archive(base_name=product_name, format='zip', base_dir=product_name)